Improved pg-build.sh.

+++ tmp/test/example1--0.0.1.sql
@@ -1,3 +1,6 @@
+-- ----------------------------------------------------
+-- example1--0.0.1.sql
+
 -- complain if script is sourced in psql, rather than via CREATE EXTENSION
 \echo Use "CREATE EXTENSION example1" to load this file. \quit
 
@@ -6,11 +9,14 @@ CREATE FUNCTION EXAMPLE1_VERSION() RETURNS varchar
     LANGUAGE C STRICT;
 
 CREATE FUNCTION cubic_poly(
-  x   double precision,
-  c0  double precision,
-  c1  double precision,
-  c2  double precision,
-  c3  double precision
-) RETURNS double precision
+  x   float8,
+  c0  float8,
+  c1  float8,
+  c2  float8,
+  c3  float8
+) RETURNS float8
     AS '$libdir/example1', 'example1_swig_cubic_poly'
     LANGUAGE C STRICT;
+
+-- ----------------------------------------------------
+
